+++ Sources/AllTestz/main.swift
@@ -12,6 +12,21 @@ protocol RxTestCase {
 }
 
 
+final class ObservableOfTypeTest_ : ObservableOfTypeTest, RxTestCase {
+    #if os(macOS)
+    required override init() {
+        super.init()
+    }
+    #endif
+
+    static var allTests: [(String, (ObservableOfTypeTest_) -> () -> ())] { return [
+    ("test_ofTypeComplete", ObservableOfTypeTest.test_ofTypeComplete),
+    ("test_ofTypeDowncastComplete", ObservableOfTypeTest.test_ofTypeDowncastComplete),
+    ("test_ofTypeNoInstanceComplete", ObservableOfTypeTest.test_ofTypeNoInstanceComplete),
+    ("test_ofTypeDisposed", ObservableOfTypeTest.test_ofTypeDisposed),
+    ] }
+}
+
 final class ObservableWithLatestFromTest_ : ObservableWithLatestFromTest, RxTestCase {
     #if os(macOS)
     required override init() {
@@ -1938,6 +1953,7 @@ func XCTMain(_ tests: [() -> ()]) {
 #endif
 
     XCTMain([
+        testCase(ObservableOfTypeTest_.allTests),
         testCase(ObservableWithLatestFromTest_.allTests),
         testCase(ObservableOptionalTest_.allTests),
         testCase(AnomaliesTest_.allTests),